requirements.txt
opencv_contrib_python==4.8.1.78
opencv_python==4.7.0.68
pandas==1.5.3
playsound==1.3.0
pyserial==3.5
